requirements.txt
streamlit==1.33.0
pypdf==4.2.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
numpy==1.26.4
groq==0.9.0